requirements.txt
# Core Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# LLM & AI Frameworks
langchain>=0.1.0
langgraph>=0.0.20
mcp>=1.0.0

# Vector Database & Embeddings
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Machine Learning & Data Science
torch>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# HTTP & Async
aiohttp>=3.9.0
requests>=2.31.0

# Data Processing & Analysis
pytrends>=4.9.0
pyyaml>=6.0

# System Monitoring & Process Management
psutil>=5.9.0
watchdog>=3.0.0

# CLI & Terminal UI
rich>=13.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Type Checking (optional but recommended)
typing-extensions>=4.8.0
